Clean up FFI code layout a little
