+++ test_rs/src/include/ffi/auth.rs
